+++ src/shortcuts/index.ts
@@ -1,6 +1,7 @@
 // Shortcuts panel module exports
 export * from './commands';
 export * from './configuration-manager';
+export * from './drag-drop-controller';
 export * from './error-handler';
 export * from './tree-data-provider';
 export * from './tree-items';